Refactored locationList.py as location_list.py
